Add new route using blueprints
