+++ src/gobstuf/blueprints/__init__.py
@@ -1,0 +1,2 @@
+from gobstuf.blueprints.hc import hc_bp
+from gobstuf.blueprints.secure import secure_bp
