+++ src/tests/auth/test_auth.py
@@ -30,27 +30,24 @@ def view_func(*args, **kwargs):
 class TestSecureRoute:
 
     def test_secure_route_allow(self, jwt_header, app):
-        route_name = "route_name"
-
         with patch("gobstuf.auth.routes.request", MockRequest()) as mock_request, \
                 patch("gobstuf.auth.routes.g", MockG()) as mock_g:
             mock_request.headers = jwt_header
-            result = secure_route("rule", view_func, name=route_name)
+            result = secure_route("rule", view_func)
 
-            assert result.__name__ == route_name
+            assert result.__name__ == view_func.__name__
+            # function needs to be called, for globals to be set
             assert result("1", "2", kwarg="3") == "access_granted_1,2_{'kwarg': '3'}"
             assert mock_g.MKS_GEBRUIKER == "test_burger"
             assert mock_g.MKS_APPLICATIE == "fp_test_burger"
 
     def test_secure_route_forbidden(self, jwt_header_forbidden, app):
-        route_name = "route_name"
-
         with patch("gobstuf.auth.routes.request", MockRequest()) as mock_request, \
                 patch("gobstuf.auth.routes.g", MockG()) as mock_g:
             mock_request.headers = jwt_header_forbidden
-            result = secure_route("rule", view_func, name=route_name)
+            result = secure_route("rule", view_func)
 
-            assert result.__name__ == route_name
+            assert result.__name__ == view_func.__name__
             assert result("1", "2", kwarg="3") == ("Forbidden", 403)
             assert mock_g.MKS_GEBRUIKER is None
             assert mock_g.MKS_APPLICATIE is None
@@ -66,10 +63,10 @@ class TestAuth(TestCase):
             view_name = 'any view'
             mock_request.scheme = 'http(s)'
             mock_request.host = 'any host'
-            mock_request.base_url = ''
+            mock_request.blueprint = 'bp'
 
             mock_url_for.return_value = '/any url'
 
             result = get_auth_url(view_name)
             self.assertEqual(result, "http(s)://any host/any url")
-            mock_url_for.assert_called_with('any view')
+            mock_url_for.assert_called_with('bp.any view')
